pyproject.toml
[project]
name = "devsentinel"
version = "0.1.0"
description = "Autonomous Docker monitoring & incident remediation system with AI-powered diagnosis and Telegram alerts"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.135.3",
    "groq>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devsentinel"]
